vendor/omni: Add chromium_prebuilt.sh
Squashed commit:
vendor/omni: Fix-ups for prebuilt Chromium
https://github.com/omnirom/android_vendor_omni/commit/9a7b0bf3367cc6ae02c5e520ea7b840e386cc924

This script will pull the necessary libs from a complete build and generate proper
makefiles in order to support using a prebuilt Chromium. Files are placed in
prebuilts/chromium/-devicenamehere-, and are not tracked by git, so they are completely
out of the way.

PS3:
-replace git -C with params compatible to old git versions

Change-Id: I3141defcbd4108178881f7203f3105aecc27378b

Conflicts:
	config/common.mk